Remove a warning when running the tool for the first time.
Only remove the symlinked udev rule if it exists.

+++ vid-pid-to-udev-rule.sh
@@ -61,5 +61,7 @@ chmod 644 $OUTFILE
 echo
 echo "Symlinking $UDEV_RULE"
 echo "to $OUTFILE"
-rm $UDEV_RULE
+if [ -e $UDEV_RULE ]; then
+	rm $UDEV_RULE
+fi
 ln -s $OUTFILE $UDEV_RULE
